+++ CHANGELOG.md
@@ -2,11 +2,10 @@
 
 Todas as principais alterações no tema serão documentadas nesse arquivo
 
-## Unreleased
-
 ## 0.0.1 - 2025-12-02
 
 -   Criação do tema e configurações básicas (cores, fontes, layout, etc)
 -   Criada página inicial customizada com seções modulares seguindo o layout de landing page.
 -   Adicionada página de configurações do tema em “Aparência → Configurações do Tema” com campos de banner, headline, contatos e webinar.
 -   Criado custom post type de depoimentos com campo adicional de Função / Cargo e integração com a seção de testimonials da home.
+-   Adicionados campos de URLs de redes sociais (LinkedIn, Facebook, Instagram e YouTube) nas configurações do tema.